requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
sentry-sdk==1.30.0
supervisor==4.2.4
gunicorn==21.2.0
